requirements.txt
ccxt==4.3.51
pynng==0.8.1
python-dotenv==1.2.1